devsecops-pulse/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
streamlit
pandas
plotly
